+++ internal/cliflags/flags.go
@@ -7,6 +7,7 @@ import (
 // Flags is a struct that holds all Aureus command-line flags.
 type Flags struct {
 	Bless bool
+	Lang  string
 }
 
 // Get returns the Aureus command-line flags.
@@ -22,5 +23,12 @@ func init() {
 		"aureus.bless",
 		false,
 		"replace (on disk) each failing assertion's expected output with its current output",
+	)
+
+	flag.StringVar(
+		&flags.Lang,
+		"aureus.lang",
+		"",
+		"only execute tests that have an input or output in the specified language",
 	)
 }